remove legacy recovery tests
