+++ test/extended/dr/common.go
@@ -8,7 +8,6 @@ import (
 	"crypto/x509"
 	"encoding/pem"
 	"fmt"
-	"os/exec"
 	"strings"
 	"text/tabwriter"
 	"time"
@@ -16,13 +15,13 @@ import (
 	g "github.com/onsi/ginkgo/v2"
 	o "github.com/onsi/gomega"
 	"github.com/openshift/library-go/test/library"
-	"github.com/openshift/origin/test/e2e/upgrade"
 	exutil "github.com/openshift/origin/test/extended/util"
 	"github.com/openshift/origin/test/extended/util/image"
 	"github.com/stretchr/objx"
 	xssh "golang.org/x/crypto/ssh"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/apimachinery/pkg/util/wait"
@@ -31,10 +30,10 @@ import (
 	applymetav1 "k8s.io/client-go/applyconfigurations/meta/v1"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
+	corev1client "k8s.io/client-go/kubernetes/typed/core/v1"
 	"k8s.io/kubernetes/test/e2e/framework"
 	e2e "k8s.io/kubernetes/test/e2e/framework"
 	e2epod "k8s.io/kubernetes/test/e2e/framework/pod"
-	e2essh "k8s.io/kubernetes/test/e2e/framework/ssh"
 	"k8s.io/kubernetes/test/utils"
 )
 
@@ -56,30 +55,6 @@ const (
 `
 )
 
-func runCommandAndRetry(command string) string {
-	const (
-		maxRetries = 10
-		pause      = 10
-	)
-	var (
-		retryCount = 0
-		out        []byte
-		err        error
-	)
-	e2e.Logf("command '%s'", command)
-	for retryCount = 0; retryCount <= maxRetries; retryCount++ {
-		out, err = exec.Command("bash", "-c", command).CombinedOutput()
-		e2e.Logf("output:\n%s", out)
-		if err == nil {
-			break
-		}
-		e2e.Logf("%v", err)
-		time.Sleep(time.Second * pause)
-	}
-	o.Expect(retryCount).NotTo(o.Equal(maxRetries + 1))
-	return string(out)
-}
-
 func masterNodes(oc *exutil.CLI) []*corev1.Node {
 	masterNodes, err := oc.AdminKubeClient().CoreV1().Nodes().List(context.Background(), metav1.ListOptions{
 		LabelSelector: "node-role.kubernetes.io/master",
@@ -93,20 +68,6 @@ func masterNodes(oc *exutil.CLI) []*corev1.Node {
 	return nodes
 }
 
-func clusterNodes(oc *exutil.CLI) (masters, workers []*corev1.Node) {
-	nodes, err := oc.AdminKubeClient().CoreV1().Nodes().List(context.Background(), metav1.ListOptions{})
-	o.Expect(err).NotTo(o.HaveOccurred())
-	for i := range nodes.Items {
-		node := &nodes.Items[i]
-		if _, ok := node.Labels["node-role.kubernetes.io/master"]; ok {
-			masters = append(masters, node)
-		} else {
-			workers = append(workers, node)
-		}
-	}
-	return
-}
-
 func internalIP(restoreNode *corev1.Node) string {
 	internalIp := restoreNode.Name
 	for _, address := range restoreNode.Status.Addresses {
@@ -115,15 +76,6 @@ func internalIP(restoreNode *corev1.Node) string {
 		}
 	}
 	return internalIp
-}
-
-func waitForMastersToUpdate(oc *exutil.CLI, mcps dynamic.NamespaceableResourceInterface) {
-	e2e.Logf("Waiting for MachineConfig master to finish rolling out")
-	err := wait.Poll(30*time.Second, 30*time.Minute, func() (done bool, err error) {
-		done, _ = upgrade.IsPoolUpdated(mcps, "master")
-		return done, nil
-	})
-	o.Expect(err).NotTo(o.HaveOccurred())
 }
 
 func waitForOperatorsToSettle() {
@@ -247,75 +199,6 @@ func condition(cv objx.Map, condition string) objx.Map {
 	return objx.Map(nil)
 }
 
-func nodeConditionStatus(conditions []corev1.NodeCondition, conditionType corev1.NodeConditionType) corev1.ConditionStatus {
-	for _, condition := range conditions {
-		if condition.Type == conditionType {
-			return condition.Status
-		}
-	}
-	return corev1.ConditionUnknown
-}
-
-func countReady(items []corev1.Node) int {
-	ready := 0
-	for _, item := range items {
-		if nodeConditionStatus(item.Status.Conditions, corev1.NodeReady) == corev1.ConditionTrue {
-			ready++
-		}
-	}
-	return ready
-}
-
-func fetchFileContents(node *corev1.Node, path string) string {
-	e2e.Logf("Fetching %s file contents from %s", path, node.Name)
-	out := execOnNodeWithOutputOrFail(node, fmt.Sprintf("cat %q", path))
-	return out.Stdout
-}
-
-// execOnNodeWithOutputOrFail executes a command via ssh against a
-// node in a poll loop to ensure reliable execution in a disrupted
-// environment. The calling test will be failed if the command cannot
-// be executed successfully before the provided timeout.
-func execOnNodeWithOutputOrFail(node *corev1.Node, cmd string) *e2essh.Result {
-	var out *e2essh.Result
-	var err error
-	waitErr := wait.PollImmediate(5*time.Second, defaultSSHTimeout, func() (bool, error) {
-		out, err = e2essh.IssueSSHCommandWithResult(context.TODO(), cmd, e2e.TestContext.Provider, node)
-		// IssueSSHCommandWithResult logs output
-		if err != nil {
-			e2e.Logf("Failed to exec cmd [%s] on node %s: %v", cmd, node.Name, err)
-		}
-		return err == nil, nil
-	})
-	o.Expect(waitErr).NotTo(o.HaveOccurred())
-	return out
-}
-
-// execOnNodeOrFail executes a command via ssh against a node in a
-// poll loop until success or timeout. The output is ignored. The
-// calling test will be failed if the command cannot be executed
-// successfully before the timeout.
-func execOnNodeOrFail(node *corev1.Node, cmd string) {
-	_ = execOnNodeWithOutputOrFail(node, cmd)
-}
-
-// sudoExecOnNodeOrFail executes a command under sudo with execOnNodeOrFail.
-func sudoExecOnNodeOrFail(node *corev1.Node, cmd string) {
-	sudoCmd := fmt.Sprintf(`sudo -i /bin/bash -cx "%s"`, cmd)
-	execOnNodeOrFail(node, sudoCmd)
-}
-
-// checkSSH repeatedly attempts to establish an ssh connection to a
-// node and fails the calling test if unable to establish the
-// connection before the default timeout.
-func checkSSH(node *corev1.Node) {
-	_ = execOnNodeWithOutputOrFail(node, "true")
-}
-
-func ssh(cmd string, node *corev1.Node) (*e2essh.Result, error) {
-	return e2essh.IssueSSHCommandWithResult(context.TODO(), cmd, e2e.TestContext.Provider, node)
-}
-
 func waitForReadyEtcdStaticPods(client kubernetes.Interface, masterCount int) {
 	g.By("Waiting for all etcd static pods to become ready")
 	waitForPodsTolerateClientTimeout(
@@ -325,6 +208,23 @@ func waitForReadyEtcdStaticPods(client kubernetes.Interface, masterCount int) {
 		masterCount,
 		40*time.Minute,
 	)
+}
+
+func waitForPodsTolerateClientTimeout(c corev1client.PodInterface, label labels.Selector, predicate func(corev1.Pod) bool, count int, timeout time.Duration) {
+	err := wait.Poll(60*time.Second, timeout, func() (bool, error) {
+		p, e := exutil.GetPodNamesByFilter(c, label, predicate)
+		if e != nil {
+			framework.Logf("Saw an error waiting for etcd pods to become available: %v", e)
+			// TODO tolerate transient etcd timeout only and fail other errors
+			return false, nil
+		}
+		if len(p) != count {
+			framework.Logf("Only %d of %d expected pods are ready", len(p), count)
+			return false, nil
+		}
+		return true, nil
+	})
+	o.Expect(err).NotTo(o.HaveOccurred())
 }
 
 // InstallSSHKeyOnControlPlaneNodes will create a new private/public ssh keypair,
